% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOptimalClustersTree.R
\name{plotOptimalClustersTree}
\alias{plotOptimalClustersTree}
\title{plotOptimalClustersTree}
\usage{
plotOptimalClustersTree(resObj)
}
\arguments{
\item{resObj}{is an object list, which is the output of FindMaxHomoOptimalPartitions function}
}
\value{
No return value, called for plotting the hierarchical tree of optimal clusters.
}
\description{
plotOptimalClustersTree is a support function for plotting the hierarchical tree of optimal clusters
from FindMaxHomoOptimalPartitions function.

The red node(s) are the optimal homogeneous clusters while the gray nodes are non-optimal clusters.
}
\examples{
# Running FindMaxHomoOptimalPartitions using simulation data
DataT<-SimpleSimulation(100,type=1)
obj<-FindMaxHomoOptimalPartitions(DataT,gamma=0.05)
# Plotting the result
plotOptimalClustersTree(obj)

}
