% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMaxHomoOptimalPartitions.R
\name{PrintOptimalClustersResult}
\alias{PrintOptimalClustersResult}
\title{PrintOptimalClustersResult}
\usage{
PrintOptimalClustersResult(resObj, selFeature = FALSE)
}
\arguments{
\item{resObj}{is an object list, which is the output of FindMaxHomoOptimalPartitions function}

\item{selFeature}{is a flag. If it is true, then the function shows the selected feature(s) of each optimal cluster.}
}
\value{
No return value, called for printing optimal clusters.
}
\description{
PrintOptimalClustersResult is a support function for printing the optimal clusters from FindMaxHomoOptimalPartitions function.
}
\examples{
# Running FindMaxHomoOptimalPartitions using simulation data
DataT<-SimpleSimulation(100,type=1)
obj<-FindMaxHomoOptimalPartitions(DataT,gamma=0.05)
# Printing the result
PrintOptimalClustersResult(obj)

}
