% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleSimulation.R
\name{SimpleSimulation}
\alias{SimpleSimulation}
\title{SimpleSimulation}
\usage{
SimpleSimulation(indvN = 10000, type = 1)
}
\arguments{
\item{indvN}{is a number of individuals per cluster}

\item{type}{is a type of simulation dataset. There are four types.}
}
\description{
SimpleSimulation is a support function for generating multiresolution datasets.
}
\examples{
# Running SimpleSimulation to generate a dataset.
DataT<-SimpleSimulation(100,type=1)


}
